attempted to get started at using multip-page stuff
