+++ package.json
@@ -10,9 +10,11 @@
     "material-ui": "^0.19.4",
     "mongoose": "^4.10.8",
     "node-fetch": "^2.0.0",
-    "react": "^15.5.4",
+    "react": "^15.6.2",
     "react-bootstrap": "^0.32.1",
-    "react-dom": "^15.5.4",
+    "react-dom": "^15.6.2",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-vertical-timeline-component": "^2.0.6"
   },
   "devDependencies": {
