+++ public/index.html
@@ -22,7 +22,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <div id="intro"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
