+++ src/routes.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { BrowserRouter, Route } from 'react-router-dom'
+import App from './app';
+import Home from './home';
+import Signup from './signUp'
+import {Switch} from 'react-router'; 
+
+export default (
+  <BrowserRouter>
+    <Switch>
+        <Route exact path="/" component={Home}/>
+        <Route exact path="/signup" component={Signup}/>
+    </Switch>
+  </BrowserRouter>
+);
