+++ src/signUp.js
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+
+export default class Signup extends Component {
+  render() {
+    return (
+      <div id="contact">
+        This is the contact me page.
+      </div>
+    );
+  }
+}
